fix(pr-review): address CodeRabbit feedback on PR #15
Server:
- admin/orgs.get: cast count(*) to ::int so memberCount/projectCount
  are JS numbers at runtime, matching the declared type.
- admin/settings.get + middleware/auth: default registration policy
  to "closed" (safer fallback when no row is configured).
- admin/settings.patch: validate all fields before any write so a 400
  never leaves a half-applied state.
- orgs/.../members/[userId].delete + transfer-ownership.post: throw
  500 on unhandled removeMembership/transferOrgOwnership reasons
  instead of falling through and returning ok.
- projects/.../company/start.post: fail fast (409) when project or
  ownerOrgId is missing instead of resolving credentials with a null
  org context.
- utils/platform-settings: introduce SettingValueByKey map so
  getSetting/setSetting are key-typed end-to-end (no more unknown
  values paired with arbitrary keys).
- utils/org-store: take FOR UPDATE lock on the org row in
  updateMembershipRole/removeMembership/transferOrgOwnership so
  concurrent admin/owner mutations on the same org serialize and
  cannot race past the last-admin / owner-immutable invariants.

Frontend:
- AnthropicOAuthCard: poll disk-state every 2s while a credential is
  connected so out-of-band changes (CLI refresh, file removed) reflect
  in the UI without a reload.
- ProjectCreateDialog: replace hardcoded workspace-selector strings
  with $t() lookups; add German translations.
- admin/settings/registration: lock the form while the initial fetch
  is pending so admins can't save the fallback default before the
  real settings arrive; fix grammar of the invite-bypass note.
- onboarding/create-org: split create-failure from refresh-failure so
  a successful POST followed by a refresh hiccup doesn't surface as
  "Could not create org" (which would prompt a duplicate retry).
- settings/orgs/[slug]/index: catch clipboard.writeText() rejections
  in copyLink() so denied permissions surface a hint instead of
  failing silently.

Docs:
- platform-admin-and-loose-ends + shadcn-ui-restructure: add language
  tags to fenced code blocks (markdownlint MD040); drop the
  generated-artifact link to .nuxt/components.d.ts.

+++ server/api/admin/orgs.get.ts
@@ -33,8 +33,8 @@ export default defineEventHandler(async (event) => {
       ownerUserId: orgs.ownerUserId,
       ownerEmail: users.email,
       createdAt: orgs.createdAt,
-      memberCount: sql<number>`(SELECT count(*) FROM ${orgMemberships} WHERE ${orgMemberships.orgId} = ${orgs.id})`,
-      projectCount: sql<number>`(SELECT count(*) FROM ${projects} WHERE ${projects.ownerOrgId} = ${orgs.id})`,
+      memberCount: sql<number>`(SELECT count(*)::int FROM ${orgMemberships} WHERE ${orgMemberships.orgId} = ${orgs.id})`,
+      projectCount: sql<number>`(SELECT count(*)::int FROM ${projects} WHERE ${projects.ownerOrgId} = ${orgs.id})`,
     })
     .from(orgs)
     .leftJoin(users, eq(users.id, orgs.ownerUserId))
